Time: 4 ms (71.73%), Space: 54.6 MB (38.12%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -12,7 +12,7 @@ class Solution {
             windowsum+=nums[i]-nums[i-k];
             maxsum=Math.max(maxsum,windowsum);
         }
-        return (double)(maxsum/k);
+        return (maxsum/k);
         
     }
 }
